+++ jans-lock/server/src/main/resources/log4j2.xml
@@ -5,7 +5,7 @@
             <PatternLayout pattern="%d{dd-MM-yyyy HH:mm:ss.SSS} %-5p %C{4} %F:%L- %m%n" />
         </Console>
         
-        <RollingFile name="FILE" fileName="${sys:log.base}/logs/jans-link.log" filePattern="${sys:log.base}/logs/jans-link-%d{yyyy-MM-dd}-%i.log">
+        <RollingFile name="FILE" fileName="${sys:log.base}/logs/jans-lock.log" filePattern="${sys:log.base}/logs/jans-lock-%d{yyyy-MM-dd}-%i.log">
             <PatternLayout pattern="%d{dd-MM-yyyy HH:mm:ss.SSS} %-5p [%t] %C{4} %F:%L- %m%n" />
 
             <Policies>
@@ -14,7 +14,7 @@
             <DefaultRolloverStrategy max="30" />
         </RollingFile>
 
-		<RollingFile name="JANS_LINK_PERSISTENCE_FILE" fileName="${sys:log.base}/logs/jans-link_persistence.log" filePattern="${sys:log.base}/logs/jans-link_persistence-%d{yyyy-MM-dd}-%i.log">
+		<RollingFile name="JANS_LOCK_PERSISTENCE_FILE" fileName="${sys:log.base}/logs/jans-lock_persistence.log" filePattern="${sys:log.base}/logs/jans-lock_persistence-%d{yyyy-MM-dd}-%i.log">
 			<PatternLayout pattern="%d %-5p [%t] [%C{6}] (%F:%L) - %m%n" />
 
 			<Policies>
@@ -24,7 +24,7 @@
 			<DefaultRolloverStrategy max="30" />
 		</RollingFile>
 
-		<RollingFile name="JANS_LINK_PERSISTENCE_DURATION_FILE" fileName="${sys:log.base}/logs/jans-link_persistence_duration.log" filePattern="${sys:log.base}/logs/jans-link_persistence_duration-%d{yyyy-MM-dd}-%i.log">
+		<RollingFile name="JANS_LOCK_PERSISTENCE_DURATION_FILE" fileName="${sys:log.base}/logs/jans-lock_persistence_duration.log" filePattern="${sys:log.base}/logs/jans-lock_persistence_duration-%d{yyyy-MM-dd}-%i.log">
 			<PatternLayout pattern="%d %-5p [%t] [%C{6}] (%F:%L) - %m%n" />
 
 			<Policies>
@@ -34,7 +34,7 @@
 			<DefaultRolloverStrategy max="30" />
 		</RollingFile>
 
-		<RollingFile name="JANS_LINK_PERSISTENCE_LDAP_STATISTICS_FILE" fileName="${sys:log.base}/logs/jans-link_persistence_ldap_statistics.log" filePattern="${sys:log.base}/logs/jans-link_persistence_ldap_statistics-%d{yyyy-MM-dd}-%i.log">
+		<RollingFile name="JANS_LOCK_PERSISTENCE_LDAP_STATISTICS_FILE" fileName="${sys:log.base}/logs/jans-lock_persistence_ldap_statistics.log" filePattern="${sys:log.base}/logs/jans-lock_persistence_ldap_statistics-%d{yyyy-MM-dd}-%i.log">
 			<PatternLayout pattern="%d %-5p [%t] [%C{6}] (%F:%L) - %m%n" />
 
 			<Policies>
@@ -44,7 +44,7 @@
 			<DefaultRolloverStrategy max="30" />
 		</RollingFile>
 
-		<RollingFile name="JANS_LINK_SCRIPT_LOG_FILE" fileName="${sys:log.base}/logs/jans-link_script.log" filePattern="${sys:log.base}/logs/jans-link_script-%d{yyyy-MM-dd}-%i.log">
+		<RollingFile name="JANS_LOCK_SCRIPT_LOG_FILE" fileName="${sys:log.base}/logs/jans-lock_script.log" filePattern="${sys:log.base}/logs/jans-lock_script-%d{yyyy-MM-dd}-%i.log">
 			<PatternLayout pattern="%d %-5p [%t] [%C{6}] (%F:%L) - %m%n" />
 
 			<Policies>
@@ -56,52 +56,52 @@
     </Appenders>
 
     <Loggers>
-        <Logger name="io.jans.link" level="${log4j.default.log.level}">
+        <Logger name="io.jans.lock" level="${log4j.default.log.level}">
             <AppenderRef ref="FILE" />
         </Logger>
 
 
 		<Logger name="io.jans.orm" level="${log4j.default.log.level}" additivity="false">
-			<AppenderRef ref="JANS_LINK_PERSISTENCE_FILE" />
+			<AppenderRef ref="JANS_LOCK_PERSISTENCE_FILE" />
 		</Logger>
 
 		<Logger name="com.unboundid.ldap.sdk.LDAPConnection" level="${log4j.default.log.level}" additivity="false">
-			<AppenderRef ref="JANS_LINK_PERSISTENCE_FILE" />
+			<AppenderRef ref="JANS_LOCK_PERSISTENCE_FILE" />
 		</Logger>
 		<logger name="com.couchbase.client" level="${log4j.default.log.level}" additivity="false">
-			<AppenderRef ref="JANS_LINK_PERSISTENCE_FILE" />
+			<AppenderRef ref="JANS_LOCK_PERSISTENCE_FILE" />
 		</logger>
 
 		<Logger name="io.jans.orm.ldap.operation.watch" level="${log4j.default.log.level}" additivity="false">
-			<AppenderRef ref="JANS_LINK_PERSISTENCE_DURATION_FILE" />
+			<AppenderRef ref="JANS_LOCK_PERSISTENCE_DURATION_FILE" />
 		</Logger>
 
 		<Logger name="io.jans.orm.couchbase.operation.watch" level="${log4j.default.log.level}" additivity="false">
-			<AppenderRef ref="JANS_LINK_PERSISTENCE_DURATION_FILE" />
+			<AppenderRef ref="JANS_LOCK_PERSISTENCE_DURATION_FILE" />
 		</Logger>
 
 		<Logger name="io.jans.orm.watch" level="${log4j.default.log.level}" additivity="false">
-			<AppenderRef ref="JANS_LINK_PERSISTENCE_DURATION_FILE" />
+			<AppenderRef ref="JANS_LOCK_PERSISTENCE_DURATION_FILE" />
 		</Logger>
 
 		<Logger name="io.jans.as.server.service.status.ldap" level="${log4j.default.log.level}" additivity="false">
-			<AppenderRef ref="JANS_LINK_PERSISTENCE_LDAP_STATISTICS_FILE" />
+			<AppenderRef ref="JANS_LOCK_PERSISTENCE_LDAP_STATISTICS_FILE" />
 		</Logger>
 
 		<Logger name="io.jans.service.PythonService" level="${log4j.default.log.level}" additivity="false">
-			<AppenderRef ref="JANS_LINK_SCRIPT_LOG_FILE" />
+			<AppenderRef ref="JANS_LOCK_SCRIPT_LOG_FILE" />
 		</Logger>
 
 		<Logger name="io.jans.service.custom.script" level="${log4j.default.log.level}" additivity="false">
-			<AppenderRef ref="JANS_LINK_SCRIPT_LOG_FILE" />
+			<AppenderRef ref="JANS_LOCK_SCRIPT_LOG_FILE" />
 		</Logger>
 
 		<Logger name="io.jans.as.server.service.custom" level="${log4j.default.log.level}" additivity="false">
-			<AppenderRef ref="JANS_LINK_SCRIPT_LOG_FILE" />
+			<AppenderRef ref="JANS_LOCK_SCRIPT_LOG_FILE" />
 		</Logger>
 		
 		<logger name="io.jans.service.external" level="${log4j.default.log.level}" additivity="false">
-			<AppenderRef ref="JANS_LINK_SCRIPT_LOG_FILE" />
+			<AppenderRef ref="JANS_LOCK_SCRIPT_LOG_FILE" />
 		</logger>
 
         <Root level="INFO">
